refactor: move main components into dedicated modules
