+++ software/tekafp/__init__.py
@@ -1,0 +1,4 @@
+from importlib.metadata import version
+
+
+__version__ = version("tek-afp")
